requirements.txt
apache-airflow==2.9.1
apache-airflow-providers-amazon==8.20.0
apache-airflow-providers-common-io==1.3.1
apache-airflow-providers-common-sql==1.12.0
apache-airflow-providers-fab==1.0.4
apache-airflow-providers-ftp==3.8.0
apache-airflow-providers-http==4.10.1
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-postgres==5.10.2
apache-airflow-providers-smtp==1.6.1

nltk==3.9.1
pandas==2.3.0
Requests==2.32.4
setuptools==66.1.1
psycopg2-binary==2.9.10
pyarrow